+++ R/environment.R
@@ -153,7 +153,8 @@ bru_log_message <- function(..., domain = NULL, appendLF = TRUE,
 #' log messages of verbosity \eqn{\le} are stored. Default: Inf, to store all messages.}
 #' \item{bru_run}{If TRUE, run inference. Otherwise only return configuration needed
 #'   to run inference.}
-#' \item{bru_max_iter}{maximum number of inla iterations, default 10. Also see the `bru_method$rel_tol` option below.}
+#' \item{bru_max_iter}{maximum number of inla iterations, default 10.
+#'  Also see the `bru_method$rel_tol` and related options below.}
 #' \item{bru_initial}{An `inla` object returned from previous calls of
 #'   `INLA::inla`, [bru()] or [lgcp()], or a list of named vectors of starting
 #'   values for the latent variables. This will be used as a
@@ -187,6 +188,8 @@ bru_log_message <- function(..., domain = NULL, appendLF = TRUE,
 #' \item{rel_tol}{Stop the iterations when the largest change in linearisation point
 #' (the conditional latent state mode) in relation to the estimated posterior
 #' standard deviation is less than `rel_tol`. Default 0.01 (one percent).}
+#' \item{max_step}{The largest allowed line search step factor. Factor 1 is the
+#' full INLA step. Default is 2.}
 #' }
 #' }
 #' \item{`inla()` options}{
@@ -283,7 +286,8 @@ bru_options_default <- function() {
       taylor = "pandemic",
       search = "all",
       factor = (1 + sqrt(5)) / 2,
-      rel_tol = 0.01
+      rel_tol = 0.01,
+      max_step = 2
     ),
     # bru_initial: NULL
     # inla options